refactor(board) : board server fetching으로 변경
